requirements.txt
numpy==1.23.5
python-decouple
beautifulsoup4==4.12.2
html5lib

aiofiles==24.1.0
httpx[http2]==0.27.2
apscheduler==3.10.4

pydantic==2.9.2
python-slugify
motor==3.6.0
pymongo==4.9.1
mongo-migrate


fastapi==0.115.0
uvicorn[standard]==0.32.0

pytest==8.3.3
pytest-asyncio==0.23.7